Keep scatter coverage within quality budgets
